prototype continuation for ejb syncthreados
