+++ dev/com.ibm.ws.webcontainer/src/com/ibm/ws/webcontainer/osgi/metadata/WebModuleMetaDataImpl.java
@@ -12,16 +12,12 @@
  *******************************************************************************/
 package com.ibm.ws.webcontainer.osgi.metadata;
 
-import java.util.List;
-
 import org.osgi.framework.ServiceRegistration;
 
 import com.ibm.websphere.csi.J2EEName;
-import com.ibm.ws.javaee.dd.common.EnvEntry;
 import com.ibm.ws.runtime.metadata.ApplicationMetaData;
 import com.ibm.ws.runtime.metadata.ComponentMetaData;
 import com.ibm.ws.runtime.metadata.MetaDataImpl;
-import com.ibm.ws.runtime.metadata.SyncToOSThreadMetaData;
 import com.ibm.ws.webcontainer.osgi.webapp.WebAppConfiguration;
 import com.ibm.wsspi.webcontainer.metadata.BaseJspComponentMetaData;
 import com.ibm.wsspi.webcontainer.metadata.WebCollaboratorComponentMetaData;
@@ -31,7 +27,7 @@ import com.ibm.wsspi.webcontainer.webapp.WebAppConfig;
 /**
  *  WAS impl also implements ComponentMetaDataFactory for dynamic creation of ComponentMetaData
  */
-public class WebModuleMetaDataImpl extends MetaDataImpl implements WebModuleMetaData, SyncToOSThreadMetaData {
+public class WebModuleMetaDataImpl extends MetaDataImpl implements WebModuleMetaData {
 
     private final ApplicationMetaData applicationMetaData;
     private J2EEName j2eeName;
@@ -184,24 +180,5 @@ public class WebModuleMetaDataImpl extends MetaDataImpl implements WebModuleMeta
     
     public void setHasHAM(boolean hamFound) {
         hasHAM = (hamFound ? Boolean.TRUE : Boolean.FALSE);
-    }
-
-    @Override
-    public boolean isSyncToOSThreadEnabled() {
-
-        boolean syncToOSThread = false;
-
-        List<EnvEntry> envEntries = webAppConfig.getEnvEntries();
-
-        // Copied from SecurityServletConfiguratorHelper.java.
-        final String SYNC_TO_OS_THREAD_ENV_ENTRY_KEY = "com.ibm.websphere.security.SyncToOSThread";
-
-        for (EnvEntry envEntry : envEntries) {
-            if (SYNC_TO_OS_THREAD_ENV_ENTRY_KEY.equals(envEntry.getName())) {
-                syncToOSThread = Boolean.parseBoolean(envEntry.getValue());
-                break;
-            }
-        }
-        return syncToOSThread;
     }
 }
